# backend/docs/swagger.yaml
definitions:
...
  internal_api_user.UserCreateContainerRequest:
    properties:
      cpu:
        type: integer
      disk:
        type: integer
      image:
        type: string
      ipv4_mapping_limit:
        type: integer
      ipv4_pool_limit:
        type: integer
      ipv6_mapping_limit:
        type: integer
      ipv6_pool_limit:
        type: integer
      memory:
        type: integer
      name:
        type: string
      password:
        type: string
      reverse_proxy_limit:
        type: integer
      traffic_limit:
        type: integer
    required:
    - image
    - name
    type: object
  lxdapi_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  models.CreateContainerRequest:
    properties:
      allow_nesting:
        type: boolean
      cpu:
        type: integer
      cpu_allowance:
        type: integer
      disk:
        type: integer
      egress:
        type: integer
      image:
        type: string
      ingress:
        type: integer
      io_read:
        type: integer
      io_write:
        type: integer
      ipv4_mapping_limit:
        type: integer
      ipv4_pool_limit:
        type: integer
      ipv6_mapping_limit:
        type: integer
      ipv6_pool_limit:
        type: integer
      memory:
        type: integer
      memory_swap:
        type: boolean
      name:
        type: string
      password:
        type: string
      privileged:
        type: boolean
      processes_limit:
        type: integer
      remark:
        type: string
      reverse_proxy_limit:
        type: integer
      traffic_limit:
        type: integer
      username:
        type: string
    required:
    - image
    - name
    type: object
...
info:
  contact: {}
paths:
...
  /api/admin/cache/containers:
    get:
      consumes:
      - application/json
      description: 从缓存获取所有容器信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取容器缓存
      tags:
      - Admin API - 缓存管理
...
  /api/admin/containers:
    get:
      consumes:
      - application/json
      description: 获取容器列表，可按用户筛选
      parameters:
      - description: 用户ID
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取容器列表
      tags:
      - Admin API - 容器管理
  /api/admin/containers/:name:
    delete:
      consumes:
      - application/json
      description: 删除指定容器
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除任务已创建
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 缺少参数
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 删除失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 删除容器
      tags:
      - Admin API - 容器管理
    get:
      consumes:
      - application/json
      description: 获取指定容器的详细信息
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 缺少参数
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取容器详情
      tags:
      - Admin API - 容器管理
...
  /api/system/containers:
    get:
      consumes:
      - application/json
      description: 获取所有容器的列表信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取容器列表
      tags:
      - System API - 容器管理
    post:
      consumes:
      - application/json
      description: 创建一个新的LXD容器，支持配置CPU、内存、磁盘、带宽等资源
      parameters:
      - description: 容器创建参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateContainerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功，返回任务ID
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 创建失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建容器
      tags:
      - System API - 容器管理
  /api/system/containers/{name}:
    delete:
      consumes:
      - application/json
      description: 删除指定名称的容器及相关数据（危险操作）
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除任务已创建
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 缺少容器名称
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 创建任务失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除容器
      tags:
      - System API - 容器管理
    get:
      consumes:
      - application/json
      description: 获取指定容器的详细信息和运行状态
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 缺少容器名称
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取容器详情
      tags:
      - System API - 容器管理
...
  /api/user/containers:
    get:
      consumes:
      - application/json
      description: 获取当前用户的所有容器
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - UserSession: []
      summary: 获取容器列表
      tags:
      - User API - 容器管理
    post:
      consumes:
      - application/json
      description: 用户创建新容器，使用管理员预设的配置参数
      parameters:
      - description: 容器创建参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_user.UserCreateContainerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功，返回任务ID
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 创建失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - UserSession: []
      summary: 创建容器
      tags:
      - User API - 容器管理
  /api/user/containers/:name:
    delete:
      consumes:
      - application/json
      description: 删除用户自己的容器
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除任务已创建
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "403":
          description: 无权操作
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 删除失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - UserSession: []
      summary: 删除容器
      tags:
      - User API - 容器管理
    get:
      consumes:
      - application/json
      description: 获取指定容器的详细信息，包含状态和系统信息
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "403":
          description: 无权访问
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - UserSession: []
      summary: 获取容器详情
      tags:
      - User API - 容器管理
...
swagger: "2.0"